add HTTP codes Errors

+++ src/main/java/org/sqli/authentification/controller/CreationController.java
@@ -1,18 +1,17 @@
 package org.sqli.authentification.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.sqli.authentification.dao.GroupRepository;
 import org.sqli.authentification.dao.UserAuthentificationRepository;
 import org.sqli.authentification.dao.auth.AuthenticationOK;
-import org.sqli.authentification.dao.auth.AuthentificationError;
-import org.sqli.authentification.dao.create.GroupDto;
+import org.sqli.authentification.dao.auth.CustomError;
 import org.sqli.authentification.dao.create.UserDto;
 import org.sqli.authentification.dao.create.UserRequest;
 import org.sqli.authentification.entitie.Group;
 import org.sqli.authentification.entitie.User;
-import org.sqli.authentification.services.AuthentificationService;
 import org.sqli.authentification.services.impl.CreationServiceImpl;
 
 import java.util.List;
@@ -39,7 +38,7 @@ public class CreationController {
         Integer id = users.size()+1;
         Optional<Group> group = groupRepository.findGroupByName(userInput.getGroup());
         if(group.isEmpty()){
-            return ResponseEntity.ok(AuthentificationError.builder().error("Group (group in input) is not valid").build());
+            return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body(CustomError.builder().error("Group (group in input) is not valid").build());
         }
         UserDto userDto = UserDto.builder().id(id)
                         .login(userInput.getLogin())
@@ -50,8 +49,7 @@ public class CreationController {
         User user =  userDto.toEntity(userDto);
 
         if(creationService.validateSavingUser(user).equals("not Valid")){
-            System.out.println(creationService.validateSavingUser(user));
-            return ResponseEntity.ok(AuthentificationError.builder().error("Login (login in input) is not valid").build());
+            return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body(CustomError.builder().error("Login (login in input) is not valid").build());
         }else{
             return ResponseEntity.ok(AuthenticationOK.builder().id(user.getId())
                     .login(user.getLogin())
@@ -62,6 +60,10 @@ public class CreationController {
 
     @DeleteMapping("/user/{login}")
     public ResponseEntity<?> deleteUser(@PathVariable("login") String login){
+
+        if(userAuthentificationRepository.findByLogin(login).isEmpty()){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(CustomError.builder().error("Login (login in input) is not found").build());
+        }
         return ResponseEntity.ok(creationService.delete(login));
     }
 }
